+++ src/main/java/com/vesperin/base/ScopeAnalyser.java
@@ -5,9 +5,8 @@ import com.vesperin.base.locations.Locations;
 import com.vesperin.base.requests.BindingRequestBySignature;
 import com.vesperin.base.requests.BindingRequestByValue;
 import com.vesperin.base.spi.BindingRequest;
-import com.vesperin.base.utils.Immutable;
-import com.vesperin.base.utils.Jdt;
-import com.vesperin.base.utils.Sets;
+import com.vesperin.utils.Immutable;
+import com.vesperin.utils.Sets;
 import com.vesperin.base.visitors.DeclarationsAfterVisitor;
 import com.vesperin.base.visitors.ScopeVisitor;
 import com.vesperin.base.visitors.StatementsSelectionVisitor;
