+++ src/main/java/com/vesperin/utils/StopWatch.java
@@ -1,4 +1,4 @@
-package com.vesperin.base.utils;
+package com.vesperin.utils;
 
 import java.util.Objects;
