+++ src/test/java/com/vesperin/base/ContextTest.java
@@ -2,7 +2,7 @@ package com.vesperin.base;
 
 import com.vesperin.base.locations.Locations;
 import com.vesperin.base.locators.UnitLocation;
-import com.vesperin.base.utils.Immutable;
+import com.vesperin.utils.Immutable;
 import org.junit.Test;
 
 import java.util.Arrays;
